Adding smarter defaults for tests and fixing command line prompt
Smarter defaults:
Defaults were hard to import in test_config since you'd need to define ALL of
them as-written. Updated so the defaults are set, and overridden with
test_config.py (a file ignored by .gitignore, though I do not recommend
storing a CAS password in here anyhow)

Fixing command-line prompt:
It was awkward when prompted for your CAS password, not knowing where your
password is being authenticated against. I added the host so at least
you can be educated as to which server you're shooting your CAS password to